Backend/requirements.txt
flask
openai
flask_cors
gunicorn
pillow
python-dotenv